+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Release 1.1.0 (TBD)
 New:
 * Add KB3 bolometer cameras. (#31)
 
+Changes:
+* Remove `/projects/cadmesh/` from default paths for rsm files. (#36)
+
 Release 1.0.0 (7 Aug 2017)
 --------------------------